+++ max_div/benchmarks/_problems/__init__.py
@@ -1,0 +1,10 @@
+"""
+Package containing the actual definitions of benchmark problems.
+"""
+
+# importing this will trigger import of all defined benchmark problems, also triggering execution
+# of their decorators and hence their registration in the benchmark problem registry
+IMPORT_ME_FOR_BENCHMARK_PROBLEM_DISCOVERY = object()
+
+# import actual benchmark problems to register them
+from ._problem_a1 import BenchmarkProblem_A1